Mediator for security_console_scanner
