+++ keycloak_scanner/scanners/security_console_scanner.py
@@ -1,61 +1,39 @@
-from typing import Dict
+from typing import Dict, List
 
 from keycloak_scanner.logging.vuln_flag import VulnFlag
-from keycloak_scanner.scanners.realm_scanner import Realms
 from keycloak_scanner.scanners.scanner import Scanner
 from keycloak_scanner.scanners.scanner_pieces import Need
-from keycloak_scanner.scanners.types import Realm
+from keycloak_scanner.scanners.types import Realm, SecurityConsole, securityConsoleType
 
 URL_PATTERN = '{}/auth/realms/{}/clients-registrations/default/security-admin-console'
 
 
-class SecurityConsoleResult:
+class SecurityConsoleScanner(Scanner[SecurityConsole]):
 
-    def __init__(self, realm: Realm, url: str, json: dict, secret: dict = None):
-        self.realm = realm
-        self.url = url
-        self.json = json
-        self.secret = secret
-
-    def __eq__(self, other):
-        if isinstance(other, SecurityConsoleResult):
-            return self.realm == other.realm and self.url == other.url and self.json == other.json and self.secret == other.secret
-        return NotImplemented
-
-    def __repr__(self):
-        return f"SecurityConsoleResult({repr(self.realm)}, '{self.url}', '{self.json}', '{self.secret}')"
-
-
-class SecurityConsoleResults(Dict[str, SecurityConsoleResult]):
-    pass
-
-
-class SecurityConsoleScanner(Need[Realms], Scanner[SecurityConsoleResult]):
     """
     TODO: replace with a client registration access scanner
     """
 
     def __init__(self, **kwargs):
-        super().__init__(**kwargs)
+        super().__init__(result_type=securityConsoleType, **kwargs)
 
-    def perform(self, realms: Realms, **kwargs) -> (SecurityConsoleResults, VulnFlag):
+    def perform(self, realm: Realm, **kwargs) -> (List[SecurityConsole], VulnFlag):
 
-        results = SecurityConsoleResults()
         vf = VulnFlag()
+        result = None
 
-        for realm in realms:
-            url = URL_PATTERN.format(super().base_url(), realm.name)
-            r = super().session().get(url)
-            if r.status_code != 200:
-                super().verbose('Bad status code for {}: {}'.format(url, r.status_code))
+        url = URL_PATTERN.format(super().base_url(), realm.name)
+        r = super().session().get(url)
+        if r.status_code != 200:
+            super().verbose('Bad status code for {}: {}'.format(url, r.status_code))
 
-            else:
-                super().find('SecurityAdminConsole', 'Find a security-admin-console {}: {}'.format(url, r.status_code))
-                results[realm.name] = SecurityConsoleResult(realm, url, r.json())
-                vf.set_vuln()
-                if 'secret' in r.json():
-                    secret = r.json()["secret"]
-                    super().info(f'find a secret in security console (realm {realm.name}) : {secret}')
-                    results[realm.name].secret = secret
+        else:
+            super().find('SecurityAdminConsole', 'Find a security-admin-console {}: {}'.format(url, r.status_code))
+            result = SecurityConsole(realm, url, r.json())
+            vf.set_vuln()
+            if 'secret' in r.json():
+                secret = r.json()["secret"]
+                super().info(f'find a secret in security console (realm {realm.name}) : {secret}')
+                result.secret = secret
 
-        return results, vf
+        return [result] if result else [], vf
